services/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-multipart
minio
python-dotenv
email-validator
